Split meta into class file to match patterns
